add cluster service
